pyproject.toml
[project]
name = "datasette-cron"
version = "0.1"
description = "Scheduled tasks and background jobs for Datasette"
readme = "README.md"
authors = [{name = "Alex Garcia"}]
license = "Apache-2.0"
classifiers=[
    "Framework :: Datasette"
]
requires-python = ">=3.10"
dependencies = [
    "datasette>=1.0a23",
    "datasette-plugin-router>=0.0.1a2",
    "datasette-vite>=0.0.1a3",
    "pydantic>=2",
    "python-ulid>=3",
    "sqlite-migrate>=0.1b0",
    "sqlite-utils>=3.38",
    "croniter",
    "python-dateutil",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff>=0.15.0",
    "time-machine",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[project.urls]
Homepage = "https://github.com/datasette/datasette-cron"
Changelog = "https://github.com/datasette/datasette-cron/releases"
Issues = "https://github.com/datasette/datasette-cron/issues"
CI = "https://github.com/datasette/datasette-cron/actions"

[project.entry-points.datasette]
cron = "datasette_cron"

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[tool.setuptools.package-data]
datasette_cron = ["static/**/*", "templates/*", "manifest.json"]

[tool.setuptools.packages.find]
exclude = ["frontend"]

[tool.ty.src]
exclude = ["tests/", "samples/", "frontend/"]

